Modify src/Lib.hs
Add the concept of user states.

A user state is one of _safe_, _at risk_, or _unknown_.

By default when we create a new user their state is _unknown_.

Parameterized the User type to give us the flexibility to lift the
concept of user states to the type level and create type aliases that
allow us to define a SafeUser, AtRiskUser, etc.

The idea is to allow us to model the state transitions at the type
level and use the type system to check our logic for us.